Time stamp update - removed mapbox
Provide time stamp that appear beside the signature when a signature box is signed and accepted. Will save and load for regular and emergency plans.

MapBox was removed as i am unable to use it for PC application. loaction information only work on mobile. May try again in the future with System.Device DLL - Also unable to get it working in unity.